requirements.txt
# AI-Powered PPT Generator - Requirements
# Free Tier Optimized

# Core PPT manipulation
python-pptx==0.6.23

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Google AI SDK (FREE)
google-generativeai==0.3.2

# Google Cloud (FREE TIER)
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
